refactor(orchestrator): move dedup planning into dedup.go
Pure move of the plan/budget logic and its tests so cache.go keeps
only the IO pieces (drain and the Dedup entry point) and the new
logic is isolated for review.
